// src/model/Categories/model_catgories.ts
export interface FailureGoType {
  failure_go_id: number;
  failure_go_title: string;
  failure_go_description: string;
}

export interface SetupGoType {
  setup_go_id: number;
  setup_go_title: string;
  setup_go_description: string;
}

export interface FailureSoType {
  failure_so_id: number;
  failure_so_title: string;
  failure_so_description: string;
}

export interface SetupSoType {
  setup_so_id: number;
  setup_so_title: string;
  setup_so_description: string;
}